apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyotp==2.9.0
python-dotenv==1.0.0
email-validator==2.1.0
alembic==1.12.1
pydantic-settings==2.1.0